pyproject.toml
[tool.poetry]
name = "wanted_pre_onboarding"
version = "0.1.0"
description = ""
authors = ["yomapi <yomapi.new@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.2"
django-rest-framework = "^0.1.0"
drf-yasg = "^1.21.4"
black = "^22.10.0"
yml = "^0.0.1"
mysqlclient = "^2.1.1"
PyYAML = "^6.0"
pytest = "^7.1.3"
pytest-django = "^4.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
